+++ src/app/_components/search-box/search-box.tsx
@@ -1,76 +1,34 @@
 'use client';
 
-import {
-  ReadonlyURLSearchParams,
-  usePathname,
-  useSearchParams,
-} from 'next/navigation';
-import { ChangeEvent, FC, useEffect, useState, useTransition } from 'react';
-import { RxCross2 } from 'react-icons/rx';
+import Link from 'next/link';
+import { FC } from 'react';
+import { BiArrowBack } from 'react-icons/bi';
 
-const createUrl = (
-  pathname: string,
-  params: URLSearchParams | ReadonlyURLSearchParams
-) => {
-  const paramsString = params.toString();
-  const queryString = `${paramsString.length ? '?' : ''}${paramsString}`;
+import { SearchInput } from '@/app/_components/search-input';
+import { Tabs } from '@/app/_components/tabs';
 
-  return `${pathname}${queryString}`;
-};
+const tabs = [
+  {
+    path: `/search/chef`,
+    tabText: 'シェフ',
+  },
+  {
+    path: `/search/recipe`,
+    tabText: 'レシピ',
+  },
+];
 
 export const SearchBox: FC = () => {
-  const searchParams = useSearchParams();
-  const path = usePathname();
-  const [isPending, startTransition] = useTransition();
-  const [text, setText] = useState('');
-
-  const newParams = new URLSearchParams(searchParams.toString());
-
-  useEffect(() => {
-    if (text !== '') {
-      newParams.set('q', text);
-    } else {
-      newParams.delete('q');
-    }
-    const timeoutId = setTimeout(() => {
-      history.replaceState({}, '', createUrl(path, newParams));
-    }, 1000);
-
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [text]);
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setText(e.target.value);
-    startTransition(async () => {
-      await new Promise((resolve) => {
-        setTimeout(resolve, 1000);
-      });
-    });
-  };
-
   return (
-    <div
-      className='flex h-full flex-1 items-center rounded-lg bg-[#EEEDEF]
-        pl-4 pr-3 focus-within:ring-2 focus-within:ring-inset focus-within:ring-gray-300'
-    >
-      <input
-        type='text'
-        name='search'
-        className='flex-1 bg-[#EEEDEF] pr-3 font-bold outline-none '
-        placeholder='シェフやレシピを検索'
-        autoComplete='off'
-        onChange={handleChange}
-        value={text}
-      />
-      {isPending ? (
-        <div className='h-5 w-5 animate-spin rounded-full border-2 border-black border-t-transparent' />
-      ) : text !== '' ? (
-        <button onClick={() => setText('')} className='w-5'>
-          <RxCross2 size='1.25rem' />
-        </button>
-      ) : null}
-    </div>
+    <>
+      <div className='mb-2 flex h-9 items-center gap-x-4 px-4'>
+        <Link href='/'>
+          <BiArrowBack size='1.25rem' color='#6F6E77' />
+        </Link>
+        <SearchInput />
+      </div>
+      <Tabs tabs={tabs} />
+      <p className='gap-0 px-4 pt-5 text-xl font-bold'>話題のレシピ</p>
+    </>
   );
 };
