+++ src/app/_components/search-input/index.tsx
@@ -1,0 +1,1 @@
+export { SearchInput } from '@/app/_components/search-input/search-input';
