+++ src/app/search/layout.tsx
@@ -1,35 +1,14 @@
-import Link from 'next/link';
-import { BiArrowBack } from 'react-icons/bi';
-
 import { SearchBox } from '@/app/_components/search-box';
-import { Tabs } from '@/app/_components/tabs';
 
 type Props = {
   children: React.ReactNode;
 };
 
-const tabs = [
-  {
-    path: `/search/chef`,
-    tabText: 'シェフ',
-  },
-  {
-    path: `/search/recipe`,
-    tabText: 'レシピ',
-  },
-];
-
 export default function Layout({ children }: Props) {
   return (
-    <main className='mx-auto flex max-w-[490px] flex-col gap-y-2 pt-2'>
-      <div className='flex h-9 items-center gap-x-4 px-4'>
-        <Link href='/'>
-          <BiArrowBack size='1.25rem' color='#6F6E77' />
-        </Link>
-        <SearchBox />
-      </div>
-      <Tabs tabs={tabs} />
-      <div className='flex flex-wrap gap-x-3 gap-y-4 px-4 pt-3'>{children}</div>
+    <main className='mx-auto flex h-auto max-w-[490px] flex-col pt-2'>
+      <SearchBox />
+      <div className='flex flex-wrap gap-x-3 gap-y-4 px-4 pt-2'>{children}</div>
     </main>
   );
 }
